Make ChannelUniqueID factory a member function of ChainState
