+++ raiden/blockchain_events_handler.py
@@ -128,8 +128,7 @@ def handle_channel_new_balance(raiden, event: Event):
     assert transaction_hash, 'A mined transaction must have the hash field'
 
     chain_state = views.state_from_raiden(raiden)
-    channel_unique_id = views.get_channel_unique_id_by_token_network_id(
-        chain_state,
+    channel_unique_id = chain_state.get_channel_unique_id_by_token_network_id(
         token_network_identifier,
         channel_identifier,
     )
@@ -179,8 +178,7 @@ def handle_channel_closed(raiden, event: Event):
     assert transaction_hash, 'A mined transaction must have the hash field'
 
     chain_state = views.state_from_raiden(raiden)
-    channel_unique_id = views.get_channel_unique_id_by_token_network_id(
-        chain_state,
+    channel_unique_id = chain_state.get_channel_unique_id_by_token_network_id(
         token_network_identifier,
         channel_identifier,
     )
@@ -216,8 +214,7 @@ def handle_channel_update_transfer(raiden, event: Event):
     assert transaction_hash, 'A mined transaction must have the hash field'
 
     chain_state = views.state_from_raiden(raiden)
-    channel_unique_id = views.get_channel_unique_id_by_token_network_id(
-        chain_state,
+    channel_unique_id = chain_state.get_channel_unique_id_by_token_network_id(
         token_network_identifier,
         channel_identifier,
     )
@@ -243,8 +240,7 @@ def handle_channel_settled(raiden, event: Event):
     assert transaction_hash, 'A mined transaction must have the hash field'
 
     chain_state = views.state_from_raiden(raiden)
-    channel_unique_id = views.get_channel_unique_id_by_token_network_id(
-        chain_state,
+    channel_unique_id = chain_state.get_channel_unique_id_by_token_network_id(
         token_network_identifier,
         channel_identifier,
     )
