+++ raiden/transfer/node.py
@@ -323,8 +323,7 @@ def subdispatch_targettask(
     events = list()
     channel_state = None
     if is_valid_subtask:
-        channel_unique_id = views.get_channel_unique_id_by_token_network_id(
-            chain_state,
+        channel_unique_id = chain_state.get_channel_unique_id_by_token_network_id(
             token_network_identifier,
             channel_identifier,
         )
